+++ lib/SPVM/Builder/include/spvm_native.h
@@ -209,7 +209,7 @@ struct spvm_env {
   int32_t (*weaken)(SPVM_ENV* env, void** object_address);
   int32_t (*isweak)(SPVM_ENV* env, void** object);
   void (*unweaken)(SPVM_ENV* env, void** object_address);
-  void* (*alloc_memory_block_zero)(SPVM_ENV* env, int64_t byte_size);
+  void* (*alloc_memory_block_zero)(SPVM_ENV* env, size_t byte_size);
   void (*free_memory_block)(SPVM_ENV* env, void* block);
   int32_t (*get_memory_blocks_count)(SPVM_ENV* env);
   void* (*get_type_name_raw)(SPVM_ENV* env, void* object);
